pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.0.3",
    "flask-sqlalchemy>=3.1.1",
    "psycopg2-binary>=2.9.9",
    "openai>=1.51.1",
    "langchain>=0.3.2",
    "requests>=2.32.3",
    "google-cloud-texttospeech>=2.17.2",
    "langchain-community>=0.3.1",
]
